+++ packages/react-querybuilder/src/utils/formatQuery/formatQuery.ts
@@ -131,7 +131,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
   let parseNumbers = false;
   let placeholderFieldName = defaultPlaceholderFieldName;
   let placeholderOperatorName = defaultPlaceholderOperatorName;
-  let quoteChar = "'";
+  let quoteValuesWith = "'";
 
   if (typeof options === 'string') {
     format = options.toLowerCase() as ExportFormat;
@@ -187,7 +187,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
     parseNumbers = !!options.parseNumbers;
     placeholderFieldName = options.placeholderFieldName ?? defaultPlaceholderFieldName;
     placeholderOperatorName = options.placeholderOperatorName ?? defaultPlaceholderOperatorName;
-    quoteChar = options.quoteValuesWith ?? "'";
+    quoteValuesWith = options.quoteValuesWith ?? "'";
   }
   if (!fallbackExpression) {
     fallbackExpression =
@@ -305,7 +305,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
             quoteFieldNamesWith,
             fieldData,
             format,
-            quoteValuesWith: quoteChar,
+            quoteValuesWith,
           });
         }
         // ...otherwise use default rule processor and pass in the value
@@ -317,7 +317,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
           quoteFieldNamesWith,
           fieldData,
           format,
-          quoteValuesWith: quoteChar,
+          quoteValuesWith,
         });
       });